# The problem

When trying to fix the `unused-import` detector problem, I noticed that the detector was kind of slow.
The profiler showed that running it on the `contracts-bedrock` took about **22 seconds** :

```
         105222726 function calls (104934133 primitive calls) in 22.206 seconds

   Ordered by: internal time
   List reduced from 7611 to 10 due to restriction <10>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
    17924    3.495    0.000   11.096    0.001 slither_core.py:204(_compute_offsets_from_thing)
20003325/20003323    1.677    0.000    2.606    0.000 {built-in method builtins.isinstance}
  7061912    1.621    0.000    2.640    0.000 source_mapping.py:111(__eq__)
 10785792    1.302    0.000    1.888    0.000 source_mapping.py:101(__hash__)
     6064    0.988    0.000    1.585    0.000 data_dependency.py:390(transitive_close_dependencies)
14421547/14420087    0.780    0.000    0.781    0.000 {built-in method builtins.hash}
  7521219    0.754    0.000    1.072    0.000 naming.py:34(__eq__)
   221780    0.642    0.000    0.780    0.000 compilation_unit.py:227(filename_lookup)
  4616924    0.494    0.000    0.929    0.000 <frozen abc>:117(__instancecheck__)
  4616924    0.433    0.000    0.435    0.000 {built-in method _abc._abc_instancecheck}
```

The main culprit here is the `__compute_offsets_from_thing` method.

I noticed that the data is stored at each offset for every mapping in `slither_core`:
```python
        self._offset_to_objects: Optional[Dict[Filename, Dict[int, Set[SourceMapping]]]] = None
        self._offset_to_references: Optional[Dict[Filename, Dict[int, Set[Source]]]] = None
        self._offset_to_implementations: Optional[Dict[Filename, Dict[int, Set[Source]]]] = None
        self._offset_to_definitions: Optional[Dict[Filename, Dict[int, Set[Source]]]] = None
```
It was also running all checks within the loop, even if they did not depend on the loop offset.

# The solution

- I deduplicated the results in the `offset_to_*` mappings and created a new mapping `offset_to_min_offset` that contains the referenced min_offset. This one is only from `int` to `Set[int]`, so it's reasonably compact and fast.
- I factorized the logic to retrieve informations from these mappings in a new method `_get_offset`
- I simplified the `SourceMapping` equality function to reduce its complexity

# The results

```
         52232380 function calls (51943791 primitive calls) in 14.138 seconds

   Ordered by: internal time
   List reduced from 7612 to 10 due to restriction <10>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
9610218/9610216    1.084    0.000    1.851    0.000 {built-in method builtins.isinstance}
     6064    1.030    0.000    1.647    0.000 data_dependency.py:390(transitive_close_dependencies)
   221780    0.674    0.000    0.819    0.000 compilation_unit.py:227(filename_lookup)
    17924    0.580    0.000    1.685    0.000 slither_core.py:199(_compute_offsets_from_thing)
   108183    0.421    0.000    1.257    0.000 crytic_compile.py:265(filename_lookup)
  3734101    0.406    0.000    0.767    0.000 <frozen abc>:117(__instancecheck__)
  3734101    0.359    0.000    0.361    0.000 {built-in method _abc._abc_instancecheck}
    11612    0.355    0.000    0.674    0.000 data_dependency.py:491(convert_to_non_ssa)
    12265    0.304    0.000    0.315    0.000 node.py:117(__init__)
    34918    0.257    0.000    0.533    0.000 ssa.py:367(is_used_later)
```

We are going down from 22 seconds to about **14s** so about 35% performance improvement. 

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- **New Features**
  - Enhanced offset mapping capabilities in the core functionality.

- **Refactor**
  - Improved internal logic for handling offsets and references.

- **Bug Fixes**
  - Fixed equality checks in source mapping to use relative filenames for more accurate comparisons.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->